Getting Web Page Running and Creating Access Tokens and Redirect Tokens
I created the web page using flask in Python and we able to get the web page running by entering in my Client ID and Client Secret. For now the web pages only have text on them. The next step is to actually pull info from the Spotify API to display on the screen.