Completing Project
I finished creating and building all methods needed for this project. I then had to modify some of those methods when building out the driver class to make sure that all of the correct information was being returned. Finally I tested creating and modifying the playlist that the algorithm created and verified that everything worked and all of the correct songs were added to my Spotify account.